Move repo helper methods to tests::repo_utils
